requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
dj-config-url==0.1.1
Django==4.2.1
django-compat==1.0.15
django-cors-headers==4.0.0
django-crontab==0.7.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
phonenumbers==8.13.13
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.2.1
typing_extensions==4.6.2
urllib3==1.26.15
yarl==1.9.2
